// src/features/roomOwner/types/weeklyTimeTypes.ts
export type WeeklyTimeTypes = {
  monday: boolean;
  tuesday: boolean;
  wednesday: boolean;
  thursday: boolean;
  friday: boolean;
  saturday: boolean;
  sunday: boolean;
  weekStartHr: string;
  weekStartMin: string;
  weekEndHr: string;
  weekEndMin: string;
  weekHalfDayOne: string;
  weekHalfDayTwo: string;
};